+++ gantry/notifiers/slack.py
@@ -21,6 +21,8 @@ class SlackNotifier(Notifier):
             text = f":warning: Workload <{workload_url}|*{workload_name}*> was preempted!"
         elif event == "canceled":
             text = f":warning: Workload <{workload_url}|*{workload_name}*> canceled!"
+        elif event == "inactive":
+            text = f":zzz: Workload <{workload_url}|*{workload_name}*> appears to be inactive!"
         elif event == "failed":
             text = f":check-failed: Workload <{workload_url}|*{workload_name}*> failed!"
         elif event == "succeeded":