+++ gantry/recipe.py
@@ -155,6 +155,7 @@ class Recipe:
         timeout: int | None = None,
         start_timeout: int | None = None,
         inactive_timeout: int | None = None,
+        inactive_soft_timeout: int | None = None,
     ) -> BeakerWorkload:
         """
         Launch an experiment on Beaker. Same as the ``gantry run`` command.
@@ -168,6 +169,7 @@ class Recipe:
             timeout=timeout,
             start_timeout=start_timeout,
             inactive_timeout=inactive_timeout,
+            inactive_soft_timeout=inactive_soft_timeout,
         )
         assert workload is not None
         return workload